@@ -6,6 +6,7 @@
     "precommit": "lint-staged",
     "dev": "next",
     "build": "next build",
+    "postbuild": "next-sitemap",
     "start": "next start",
     "export": "NEXT_EXPORT=true next build && NEXT_EXPORT=true next export",
     "deploy": "yarn export && firebase deploy",
@@ -24,50 +25,48 @@
     ]
   },
   "dependencies": {
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/styles": "^4.10.0",
-    "framer-motion": "^2.9.4",
-    "gsap": "^3.5.1",
-    "isomorphic-unfetch": "^3.1.0",
-    "lax.js": "^1.2.5",
-    "next": "^10.0.1",
-    "next-pwa": "^3.1.5",
-    "next-seo": "^4.17.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-error-boundary": "^3.0.2",
-    "react-intersection-observer": "^8.30.2",
+    "@emotion/cache": "^11.10.5",
+    "@emotion/react": "^11.10.5",
+    "@emotion/server": "^11.10.0",
+    "@emotion/styled": "^11.10.5",
+    "@marsidev/react-turnstile": "^0.0.7",
+    "@mui/icons-material": "^5.11.0",
+    "@mui/material": "^5.11.8",
+    "framer-motion": "^9.0.2",
+    "lax.js": "^2.0.3",
+    "next": "^13.1.6",
+    "next-pwa": "^5.6.0",
+    "next-seo": "^5.15.0",
+    "next-sitemap": "^3.1.50",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-error-boundary": "^3.1.4",
+    "react-intersection-observer": "^9.4.1",
     "react-overdrive": "^0.0.10",
-    "react-plx": "^1.3.16",
+    "react-plx": "^2.0.1",
     "react-progressive-image": "^0.6.0",
-    "react-query": "^2.26.2",
-    "react-spring": "^8.0.27",
-    "react-transition-group": "^4.4.1",
-    "styled-components": "^5.2.1",
+    "react-query": "^3.39.3",
+    "styled-components": "^5.3.6",
     "styled-media-query": "^2.1.2",
-    "typed.js": "^2.0.11",
-    "use-axios-client": "^2.0.0",
-    "vconsole": "^3.3.4"
+    "typed.js": "^2.0.12",
+    "vconsole": "^3.15.0"
   },
   "devDependencies": {
-    "@types/material-ui": "^0.21.8",
+    "@types/material-ui": "^0.21.12",
     "@types/next": "^9.0.0",
-    "@types/node": "^14.14.7",
-    "@types/react": "^16.9.56",
-    "@types/react-dom": "^16.9.9",
-    "@types/react-transition-group": "^4.4.0",
-    "@types/styled-components": "^5.1.4",
-    "@types/styled-jsx": "^2.2.8",
-    "@zeit/next-bundle-analyzer": "^0.1.2",
-    "babel-plugin-styled-components": "^1.11.1",
-    "husky": "^4.3.0",
-    "lint-staged": "^10.5.1",
-    "next-compose-plugins": "^2.2.1",
-    "prettier": "^2.1.2",
-    "typescript": "^4.0.5"
+    "@types/node": "^18.13.0",
+    "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.0.10",
+    "@types/styled-components": "^5.1.26",
+    "babel-plugin-styled-components": "^2.0.7",
+    "eslint": "^8.33.0",
+    "eslint-config-next": "^13.1.6",
+    "husky": "^8.0.3",
+    "lint-staged": "^13.1.1",
+    "prettier": "^2.8.4",
+    "typescript": "^4.9.5"
   },
   "engines": {
-    "node": "^12.x"
+    "node": "^18.x"
   }
 }
