@@ -1,20 +1,17 @@
-import { createMuiTheme } from '@material-ui/core/styles'
-import blue from '@material-ui/core/colors/blue'
-import pink from '@material-ui/core/colors/pink'
-import red from '@material-ui/core/colors/red'
-
+import { createTheme } from '@mui/material/styles'
+import { blue, pink, red } from '@mui/material/colors'
 // Create a theme instance.
-const theme = createMuiTheme({
+const theme = createTheme({
   palette: {
     primary: blue,
     secondary: pink,
     error: {
-      main: red.A400
+      main: red.A400,
     },
     background: {
-      default: '#fff'
-    }
-  }
+      default: '#fff',
+    },
+  },
 })
 
 export default theme
